Initial Release
The problem with edge update has solved in the "sub-copy" function. / LogiBoost and BrownBoost skeleton has added.

The elapsed time is calculated from the beginning of the boosting. It doesn't include the loading of the arff and shyp files.

BanditSingleStumpLearner serialization has implemented. The rewards and the history of the pulled arms is stored in the shyp file. In the GenericBanditAlgorithm the Serialization interface is implemented, but it isn't used in the BanditSingleStumpLearner, because the rewards and the arms has been pulled need to be stored in a separate file.

The new version of decision trees has renamed as TreeLearner instead of TreeLearner3.

the code of TreeLearner was commented a little bit. The code must be polish.

some unimportant warning have been eliminated.

GenericBanditAlgorithm.h serializationFlag problem fixed

BanditProductLearner kesz es a BanditTreeLearner is.

FilterBoost v1.0