Fast resume bug (in AdaBoostMHLearner::updateWeight) fixed. The normalization was not done properly. Now it's also working in log domain so numerically it is more stable
Added README in markdown format

Bug fix: --cmatrix and --cmatrixfile confusion. Now it's all --cmatrix with 2 mandatory arguments and one optional argument which is the output file.

Minor fix: The order of --test arguments was not correct in the help section.

cmatrix command line fix for VJCascadeLearner.

with --outputinfo <filename>, the header is now output separately in <filename>.header.
"tfr" output code is now "roc" (It outputs the output of the true positive and false positive rates).
A minor change in AdaBoostMHLearner's screen output.

- A bit more information in column descriptions
- outputDescription() is no longer a pure virtual function. It becomes then optional.

OutputInfo now appends the results file instead of overwriting it.

Bug corrected in ProductLearner so now it also works for ternary base
classifiers (which can abstain by outputting 0). It was implemented but never
tested, and there was a bad bug in it: the labels were not initialized in line
157 before reconstructing them from scratch.

An optional parameter is also added:

--loop

If set, we will continue looping over the base classifiers after the first
<numBaseLearners> iterations until the edge stops increasing (as specified in
the pseudocode). In practice it rarely improves the edge, and it costs
additional computation, and it is especially expensive for ternary
classifiers. The default is false.

some minor reindentations (my fault)

- The time column in the output file starts from the last value in the existing output file (when --resume is used).
- OutputInfo clean up.
- The SoftCascade Learner doesn't not output posteriors anymore during the learning.

Conflicts:
	src/IO/OutputInfo.cpp

	modified:   src/Classifiers/AdaBoostMHClassifier.cpp
	modified:   src/IO/OutputInfo.cpp
	modified:   src/IO/OutputInfo.h
	modified:   src/StrongLearners/AdaBoostMHLearner.cpp
	modified:   src/StrongLearners/FilterBoostLearner.cpp
	modified:   src/StrongLearners/SoftCascadeLearner.cpp

New arff parser with multclass implemented as "class*" prefixed attribute names. Old simple/extended labels stay compatible.

Working, but still undecided about initweightpolicies

small compilation bug with _outStream.tellg() != 0 in OutputInfo.cpp (should compare with eof())

New ArffParser validated, sparse->dense onthefly working for Stump now.

WIP new SparseStump, getMostFrequentValuePerFeature() in InputData to give the most used feature value for each feature column.

Moved the definition current version of MB to the CMakeLists.txt

Slightly better version handling, now stored in a "version" file

Cleaned up a bit please see in the gist :

https://gist.github.com/4586305

Library compilation of multiboost

some problem with registrar methods, looks like there is no portable way to do that across library, see : http://stackoverflow.com/questions/9459980/c-global-variable-not-initialized-when-linked-through-static-libraries-but-ok

Found the right option for library compilation (just a bunch of objects)

Non-intrusive way of registering learners with cmake

All REGISTER calls are now centralized in Registrations.h. Include this file for all your modules and linked executables

Reincluding .h in CMake depends, so that they appears in Xcode/Visual Studio

Version 1.1.07